Improved condition scope with for loop support
Does not yet manage to find the condition for a C-style loop like

```
for ( my $i = 0; $i < 10; $i += 1 ) {}
```

when the cursor is at the keyword at the start of the line.

Also, the tree-sitter-perl fails processing for loops that assign to the
default variable, so can't add all the tests for that yet. Example:

```
for ( 1 .. 10 ) { say "$_" }
```

Taken a note to try and fix that or raise issues for.