`PositionTarget`: consider minimum line indentation when pouring multiline target (#1405)
We have functionality that computes minimum indentation when doing a
"pour", but we don't use it if the delimiter is `\n` because we switch
to using the editor's newline insertion to handle things like indenting
after a line ending in `:` in Python. However, if the user is explicitly
pouring something that is multiple lines, then I argue that it is safe
to do our own insertion. This change enables things like properly
dedenting after an if statement in Python

## Checklist

- [x] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [ ] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [ ] I have not broken the cheatsheet