replace treesitter #start-position with .startof
This particular query pattern is going to be very common. 
It's worth having syntactic sugar for. 
In particular, this reduces the level of indentation 
required to express that a node's associated scope
(iteration, trailing, leading, etc.)
should start or end at the start or end of a different node.

While we're here, make a related error message more helpful.