Fix E2BIG error in pre-commit (#1712)
When pre-commit runs a command, it tries to avoid putting too many
arguments on the command-line by splitting up the arg list. To do so, it
needs to take into account the number of arguments that are on the
command itself, in addition to the args it will add for the files it
needs to process. If the command in `.pre-commit-config.yaml` calls a
subcommand that has more arguments, then pre-commit's calculation of how
many args it needs to leave for the call itself will be wrong, and we
get E2BIG.

The hack is to just inline whatever subcommand gets called into
`.pre-commit-config.yaml`. Quite annoying and duplicative, but it
appears there's no way around it. See
https://github.com/pre-commit/pre-commit/issues/2841

## Checklist

- [ ] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [ ] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [ ] I have not broken the cheatsheet