migrate Go scopes statement, comment, string, and textFragment (#1685)
This begins the Go migration to treesitter.

In addition to migrating scopes, this commit includes two bug fixes.

(1)

It adds a statement scope for expression statements.
This fixes #1592.

Note that block statements are omitted.
There are many block statements in a normal Go program,
most of which are not of interest.

For example, in:

if x() {
    // nada
}

there are two statements, one that is the entire if statement,
and one block statement that runs from brace to brace.
As the typical block statement includes other statements,
and is usually part of other named statements,
it will not typically be helpful during editing.

(2)

It fixes an existing bug in the handling of paired delimiters
when used with treesitter-based scopes.

When looking for paired delimiters from inside a string,
we first search inside the string only,
then outside the string only.

This lets us find the parens in both

"(hello)"

and

("hello")

given command "take round odd".

However, that fallback logic did not get replicated
to the treesitter-based scope flow.
As a result, we never searched outside the string.

Co-authored-by: Pokey Rule <755842+pokey@users.noreply.github.com>



## Checklist

- [x] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [x] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [x] I have not broken the cheatsheet

Co-authored-by: Pokey Rule <755842+pokey@users.noreply.github.com>