stop early when there are no candidate hats for a grapheme (#1692)
No functional changes.

However it prevents a bunch of pointless runs through the
scoring functions with an empty set of candidates.

(All the early stopping conditions check for a
unique winner, not the empty set.
Those aren't worth changing, when we can trivially
avoid calling the function at all.)

The main benefit is reduced noise when logging
inside the scoring functions when debugging.



## Checklist

- [/] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [/] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [/] I have not broken the cheatsheet