Bump docusaurus to `3.0.0-alpha.0` (#1745)
We have some security warnings on a couple packages that are
dependencies of docusaurus, so I decided to upgrade to the latest to
remove those dependencies. Docusaurus uses their nightly version on
their own docs, and things seem to work for us on this version with no
issue, so I think it's ok. In addition, we were violating their peer
dependencies by using react 18; we now no longer need to do that. Also,
this version allows us to use mdx v2, which has a lot of niceties that
we'll probably want as we do our big docs refactor
(https://github.com/cursorless-dev/cursorless/issues/867)

- depends on https://github.com/cursorless-dev/cursorless/pull/1750 to
fix typedoc failure

## Checklist

- [ ] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [ ] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [ ] I have not broken the cheatsheet