optimize getTokenRemainingHatCandidates
Object.entries + map allocates a copy of the list twice.
And it ends up calling a tiny function over and over.

Using a boring nested for loop and an array
cuts hat allocation time by about 35%.

Removing items from an array one at a time is quadratic (triangular).
However, N is capped at the number of hats a user has (per grapheme),
which is fundamentally capped at a fairly small number.
And copying an array is typically a highly optimized operation.

This bit of hot code showed up in a profile from user saidelike on Slack.