Fix text fragment extractor fallbacks (#1863)
The fallback logic for determining whether to use text-based surrounding
pairs or parse-tree-based was broken for text fragment extractors based
on next-gen scope handlers (ie using the new `@textFragment` tag in a
query file).

This breakage meant that in a string like `"""hello"""` in Python, the
surrounding pair finder fell back to text-based, which resulted in it
thinking `"hello"` was a string, rather than `"""hello"""`, as would
happen if it fell back to parse-tree-based

This PR fixes the issue by unifying the fallback logic between legacy
and next-gen text fragment extractors, so that they don't fall out of
sync again

See also
https://github.com/cursorless-dev/cursorless/issues/1812#issuecomment-1691493746;
arguably that will make this PR irrelevant, but until then, it's better
to have string not be broken when we migrate a language to use
`@textFragment`

Note that the tests in this PR don't actually test the code yet, because
Python is still using legacy text fragment extractors. The tests will
start biting when this PR is merged into main and then merged into
https://github.com/cursorless-dev/cursorless/pull/1862

- This PR is required by
https://github.com/cursorless-dev/cursorless/pull/1862

## Checklist

- [ ] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [ ] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [ ] I have not broken the cheatsheet

---------

Co-authored-by: Andreas Arvidsson <andreas.arvidsson87@gmail.com>