Migrate Typescript "type" to next-gen; fix value removal (#1925)
Note that the first commit moves `typescript.scm` to
`typescript.core.scm`, because we now have some patterns that shouldn't
be defined for TSX (cast assertions eg `<foo>bar`). I made it its own
commit because otherwise the diff is confusing. Might be easier to
review
https://github.com/cursorless-dev/cursorless/pull/1925/files/dbb4e5ea4135dad6bcff272f6bc85752a650013e..a2ced3b3ad14a27b4d47d344ddacfebba65c0915,
which excludes that commit

## Checklist

- [x] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [-] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [-] I have not broken the cheatsheet