Don't focus search editor (#2091)
Partially addresses #1722

The problem is that when you focus the search editor focus is brought
back to the import/query box. For now just don't focus the editor
actually makes it so if the search editor is already in focus you can
now use Cursorless in it. The simplest way to get to the search editor
is probably use `search.action.focusNextSearchResult`. Unfortunately
there is no command to focus the results generally. Focusing the next
result we of course don't want to issue every time we do a Cursorless
command.

## Checklist

- [-] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [-] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [-] I have not broken the cheatsheet