Migrate collection item and argument scopes (#2081)
Note that we are here just migrating markdown and yaml collection. For
languages like typescript I propose that the insertion delimiter is
either `, ` for an inline collection or `,\n` for an vertical
collection. When the edit is constructed by our destination `\n` gets
replaced by `{indent}\n` or `\n{indent}` depending if it's before or
after. In short I don't think we actually should use leading or trailing
for the insertion delimiters and the destination can take care of that
on its own.

To get the correct insertion delimiter we could have a conditional
insertion delimiter predicate that actually uses the collection and not
the item itself to determine if it's a vertical collection/insertion
delimiter.

`insertionDelimiter = @list.range.isSingleLine ? ", " : ",\n"`
```
(
  (array
    (_) @collectionItem
  ) @list
  (#conditional-insertion-delimiter! @collectionItem @list ", " ",\n")
 ) 
```

Edit: Just went ahead and migrated typescript argument using the above
predicate. Personally I think it turned out quite nicely.

Will fix the second task
https://github.com/cursorless-dev/cursorless/issues/585

## Checklist

- [x] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [-] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [-] I have not broken the cheatsheet

---------

Co-authored-by: Pokey Rule <755842+pokey@users.noreply.github.com>