start parsing things from the extension side
we have 2 commands (getContent and setupStep) respectively to get the full content of the tutorial and to ask the extension to set up the state in vscode for the current step.

we already support parsing the spoken form for the command to be said by the user during the getContent command and spawning the contents in vscode for the setupStep.

We need to parse the other spoken forms and to initialize better the vscode content with the hats, etc.