support hat strokes as a hidden feature (#1810)
To use, add e.g. `-#aa10ff` after any color in settings to add `#aa10ff`
as the stroke color.

For example, my `userColor2` is `#000000-#ffffff`, and my green is
`#14ff07-#00a336`.

This is intentionally undocumented: It is a hidden feature for folks to
play around with, but only if they are paying a lot of attention, at
which point they will know that is not officially supported. :)

If we decide we like it, we can eventually see about documenting and
testing and generally "unhiding" it.

## Checklist

- [-] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [-] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [-] I have not broken the cheatsheet