Stop using `backspace` for delete
This broke when we reworked our keyboard actions setup. It's also not properly integrated with new keyboard parser. If we wanted to support special keys like `backspace` with our keyboard parser, we should define a VSCode command that sends a keyboard token to our parser, and map `backspace` to that