Stop using `backspace` for `remove` action; tweak maps (#2202)
This broke when we reworked our keyboard actions setup. It's also not
properly integrated with new keyboard parser. If we wanted to support
special keys like `backspace` with our keyboard parser, we should define
a VSCode command that sends a keyboard token to our parser, and map
`backspace` to that

We also add `remove` to default config, and tweak some maps while we're
there

## Checklist

- [-] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [x] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [-] I have not broken the cheatsheet