Fix FontMeasurementsImpl to work in a web environment (#2260)
FontMeasurementsImpl currently uses an inline script to compute font
size ratios which is not compatible with a web environment due to
Content Security Policy (CSP) rules. This PR switches to a separate
script file and adds a CSP rule for it according to the instructions in
https://code.visualstudio.com/api/extension-guides/webview#content-security-policy.

To test, change the `Editor: Font Family` setting (you can just append
`, x` to the end or make any other change). Success is seeing a tab
flash and disappear; failure is the tab showing "Loading Cursorless..."
remaining open.

---------

Co-authored-by: Barry Jaspan <bjaspan@google.com>
Co-authored-by: pre-commit-ci-lite[bot] <117423508+pre-commit-ci-lite[bot]@users.noreply.github.com>