Add a page about visual impairments to the docs (#2258)
Fixes https://github.com/cursorless-dev/cursorless/issues/935

I specifically addressed everything in that issue. If desired, I could
reference my sight free repository or any of the discussions Pokey and I
had with Parham. I wasn't entirely sure since on one hand what I wrote
up doesn't discuss anything about blindness or visual impairments that
affect the ability to see the hats regardless of size, but on the other
hand that is relatively specific and would require referencing another
package outside the repository which I don't want to cause issues.

I continue to be interested generally about ways to make cursorless
better for non-visual use but realistically this might be more of a
stretch goal and not for this PR

## Checklist

(First take in checklist not relevant since no code is changed)
- [x] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [x] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [x] I have not broken the cheatsheet

---------

Co-authored-by: Pokey Rule <755842+pokey@users.noreply.github.com>