enforce new selection for "remove" action (#2285)
## Checklist

- [ ] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [ ] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [ ] I have not broken the cheatsheet

From debugging the vscode cursorless extension, it seems the selection
is automatically updated by vscode itself when dealing with the "remove"
action while calling editBuilder.delete() (TextEditorEdit).

However it is not the case for neovim so it was using the old selection.

I fixed it so cursorless always enforce selection for that "remove"
action. Tested on neovim and vscode.

---------

Co-authored-by: Cedric Halbronn <cedric.halbronn@nccgroup.com>