Replace path with pathe (#2481)
This is a simple way of getting rid of node path references in common
and engine. I thought it best to also replace it in vscode so we don't
have different implementations for path in the bundle. I guess for
things like running tests and ours setup scrips then just using node
path is probably fine? I could go either way.

https://www.npmjs.com/package/pathe

## Checklist

- [/] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [/] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [/] I have not broken the cheatsheet