Move node dependencies out of commons and engine (#2500)
This is all about getting node out of commons. A lot of our file system
interaction had to do with running tests and reading/writing test
fixtures so I created a separate node test package.

To avoid circular dependencies I had to move a bunch of types from the
engine to commons. This make it so we now have changes in imports in a
bunch of different files. Most of the changes are quite mechanical or
boilerplate. The only actual new code is the two storage providers to
`ScopeTestRecorder` and `TestCaseRecorder`. The latter one is quite
lacking in implementation. This is just about unblocking cursorless
everywhere. A better abstraction is something we can workshop together.

And before you ask: no I can't split this into smaller pr. It feels like
a small miracle that I could solve all of the circular dependency
problems I had already :)

I wonder if we in a follow up want to move everything test related to
the test package? Generally I prefer to have test as close to the source
code as possible but in this case it's getting problematic to keeping
track of all the dependencies. If we had a test package we could make
sure that a lot of the dependencies only existed in that package and we
could add lint rules preventing them in the other packages.

## Checklist

- [/] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [/] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [/] I have not broken the cheatsheet

---------

Co-authored-by: pre-commit-ci-lite[bot] <117423508+pre-commit-ci-lite[bot]@users.noreply.github.com>
Co-authored-by: Pokey Rule <755842+pokey@users.noreply.github.com>