Remove sbd dependency (#2515)
Note that this is still `sbd`. The reason for this PR is to get rid of
the html/postcss dependences that breaks non node installations. We
didn't even have that activated so for us this doesn't break anything

My intent is to replace the implementation with something better in the
future, but for now this is just sbd with the html support removed. For
that future implementation it's nice to already have the sbd tests in
place.


## Checklist

- [x] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [/] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-taln/src/cheatsheet)
- [/] I have not broken the cheatsheet

---------

Co-authored-by: Pokey Rule <755842+pokey@users.noreply.github.com>