Extend ide capabilities (#2512)
When the ide don't have capabilities/own implementation of `indentLine`,
`outdentLine` and `insertLineAfter` fallback to doing an editor edit
instead. This makes it so that in a non code editor environment like the
web browser we still get these three actions implemented.

Ideally we would test this, but that would involve running tests with
mocked capabilities. I think that is a lot of work?

## Checklist

- [ ] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [/] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [/] I have not broken the cheatsheet

---------

Co-authored-by: Pokey Rule <755842+pokey@users.noreply.github.com>