New interface for performing edits and updating ranges (#2551)
This introduces a new interface to perform updates. The requirements I
had was:
1. Support both selections and ranges
2. Support both edits and callback
3. Support setting range expansion behavior per range/selection
4. Support updating selection internally

What this pr doesn't try to do:
1. Improve the range updater logic

Example
```ts
const {
  selections: [updatedContentSelections],
  ranges: [updatedEditRanges]
} = await new EditsUpdater(rangeUpdater, editableEditor, edits)
  .selections(contentSelections)
  .ranges(editRanges, RangeExpansionBehavior.openOpen)
  .updateEditorSelections()
  .run();
```

## Ponderings
1. Should updating editor selections be opt out instead?
1. Here I'm using two separate classes `EditsUpdater` and
`CallbackUpdater`. I'm a bit tempted to just use a single class and a
polymorphic argument. `editsOrCallback: Edit[] | () => Promise<void>`.
Something like that.
2. Strict necessary we don't actually need the two separate return
lists. Internally the range updater works with selections so we
converting ranges to selections internally. We could just return a
single list of selections and since all selections are also ranges I
don't think any of the actions would actually care?

If we embrace this polymorphic nature it could look something like this:
```ts
const [updatedEditSelections, updatedContentSelections] =
  await new EditsUpdater(rangeUpdater, editableEditor, edits)
    .add(contentSelections)
    .add(editRanges, RangeExpansionBehavior.openOpen)
    .updateEditorSelections()
    .run();
```

Personally I am a fan of the above since it became cleaner and simpler,
but I know that @pokey generally doesn't prefer union arguments if it
can be avoided.

We could of course get rid of `new` if we wanted to make it even
cleaner.

```ts
const [updatedEditSelections, updatedContentSelections] =
  await performUpdates(rangeUpdater, editableEditor, edits)
    .add(contentSelections)
    .add(editRanges, RangeExpansionBehavior.openOpen)
    .updateEditorSelections()
    .run();
```

Related to #729

## Checklist

- [/] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [/] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [/] I have not broken the cheatsheet

---------

Co-authored-by: Pokey Rule <755842+pokey@users.noreply.github.com>