fix(nix): add pkgs to fix broken devshell
I'm not sure what exactly changed, but a few things recently broke on nvim-talon, so tasks that were run when doing
neovim test were failing. Also node debug attaching broke because it wasn't finding ps, so I've explictly added.

This is a bit unfortunate because we purposefully dropped nodejs in order to use corepack to do pnpm building, but now
something is explicitly calling node, which corepack doesn't install. So to fix that I've just added nodejs package. The
caveat is that atm I'm not using a version overlay, so atm this will be 20.12.2. I think this is okay because it's not
using this to actually build things, but I can go back to doing the 18.x overlay if needed.