Added literal mark (#2593)
I'm not sure this is a direction we want to go or if this is a good
implementation, but for editors where we don't have hats it would be
nice if you could just speak the literal text instead.
`"take hello world"`
`"chuck state get value"`

In this implementation I'm only looking in the viewport and I'm solving
ambiguity by selecting the match closest to a cursor selection.

## Checklist

- [x] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [/] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [/] I have not broken the cheatsheet

- [x] Add list for literal prefix/keyword. Empty by default. test that
this behaves properly
- [x] Tag to remove prefix and have literal as a replacement for hats.
Check dfa compilation
- [x] Tag to default to preferred scope instead of containing scope

---------

Co-authored-by: pre-commit-ci-lite[bot] <117423508+pre-commit-ci-lite[bot]@users.noreply.github.com>
Co-authored-by: Phil Cohen <phillip@phillip.io>