Change default head/tail scope to be line bounded by matching pairs (#2652)
If no scope is specified for head/tail instead of defaulting to just
line we now use the smallest of line or surrounding pair interior. This
means that `"take head"` would be bounded by the surrounding pair just
like `"short paint"`. If there is no surrounding pair the line is used
just like today. You can also override this behavior by specifying line.
`"take head line"` would always use the line regardless if you are in a
surrounding pair or not.

Fixes #2434

The issue mentions a tag for this. I guess since this change is fully
extension side it would be a vscode setting, but in tempted to just have
this be enabled by default since it's a really useful feature.

## Checklist

- [x] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [/] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [/] I have not broken the cheatsheet

---------

Co-authored-by: Phil Cohen <phillip@phillip.io>