Support arbitrary scope and insertion delimiter for drop, float and puff (#2646)
This new implementation allows us to do things like `"puff token"` to
insert spaces. `"float item"` to insert `, ` etc. A bit like
https://github.com/cursorless-dev/cursorless/pull/2651.

This new implementation also fixed a few bugs around where the selection
and highlights ended up.

Note that there is still a bug with the range updater when you have an
empty selection, but that is out of scope for this pr.
https://github.com/cursorless-dev/cursorless/issues/729

Fixes  #742


## Checklist

- [x] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [/] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [/] I have not broken the cheatsheet

---------

Co-authored-by: Phil Cohen <phillip@phillip.io>