Default value for preferred scope setting (#2687)
I got unreliable behavior when we didn't have a default value here.
Turns out that if you don't have a default value get will not actually
return none. It will return a `NoValueType` object and that is evaluated
to true in the if statements so everybody probably has preferred scope
on now. This might be a change to Talon because I can't believe we
didn't catch this earlier.

## Checklist

- [/] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [/] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [/] I have not broken the cheatsheet