Added support for browser rpc (#2704)
When using Cursorless everywhere in Talon js the accessibility is not
good enough for browsers. I'm working on my own browser extension to
mitigate this.

What this pull request is doing is adding a new Cursorless everywhere
context. When the new tag is set Cursorless everywhere will use the
command client instead of access ability.

## Checklist

- [/] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [/] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [/] I have not broken the cheatsheet