Use narrow search range instead of full document for Tree sitter queries (#2714)
Today we generate all captures in the entire document whenever we do a
Tree sitter query. This is of course quite expensive for large documents
and it would be better if we could narrow the search range. That is
what's implemented in this pull request.

I also found some incorrect scm patterns that broke with this
implementation that I have fixed.

Fixes #1769

## Checklist

- [/] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [/] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [/] I have not broken the cheatsheet

---------

Co-authored-by: Phil Cohen <phillip@phillip.io>