Clean up one of scope handler (#2792)
Passing in the scope type is only necessary if you're creating the scope
handlers from that scope type. In the case where you're creating a one
of handler from a list of scope handlers. Passing in the scope type as
well just makes the code more verbose for no reason.

## Checklist

- [/] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [/] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [/] I have not broken the cheatsheet

---------

Co-authored-by: Phil Cohen <phillip@phillip.io>