Disable expand to containing token on fallback (#2798)
The fallback code we have tells Talon to expand an empty selection to
the containing token. Doing this in the search widget works fine but the
little cursor dance we do to select a token completely messes up the
file explorer when you're trying to copy or paste a file.

https://github.com/cursorless-dev/cursorless/issues/2800

## Checklist

- [/] I have added
[tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [/] I have updated the
[docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and
[cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [/] I have not broken the cheatsheet

---------

Co-authored-by: pre-commit-ci-lite[bot] <117423508+pre-commit-ci-lite[bot]@users.noreply.github.com>
Co-authored-by: Phil Cohen <phillip@phillip.io>