Log fail test to file went running tests locally (#2809)
The issue suggested showing a pop-up, but this wasn't optimal. It was
either blocking the tests from terminating or it closed immediately by
itself. I also don't want to clobber the clipboard all the time.

Let's log to a file because it seems like the best of the options.

Fixes #2805

---------

Co-authored-by: Phil Cohen <phillip@phillip.io>