Added Cursorless icon to contributed views (#2920)
Right now we get warnings about the icon's missing. Note that these
icons are only used if you drag the view away from our Cursorless side
panel and peanut as its own panel in the side bar. This is very unlikely
and this is mostly to get rid of the warning