Add sandbox setup cmd for VSCodium  (#2989)
This implicitly requires `vscode-tree-sitter-query` to be published to
the Open VSX Registry to not cause issues when running tests, as rn that
plugin is commented out in
`packages/common/src/extensionDependencies.ts` due to it not being
present there. This PR will therefor need to be a draft until that
plugin is uploaded to Open VSX; I'll probably also add to the
documentation to note the differences in commands between VSCodium and
VSCode at some point soon. There's also probably a cleaner way to do
this with a variable passed to
`packages/cursorless-vscode/src/scripts/initLaunchSandbox.ts` to swap
between vscode and vscodium mode or something, but I'll look into that
later as well.

---------

Co-authored-by: Andreas Arvidsson <andreas.arvidsson87@gmail.com>